add AbstractObjectHandler
